requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==3.2.14
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
psycopg2==2.9.3
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.3.0
whitenoise==6.2.0
